requirements.txt
joblib==0.13.2
lightgbm==2.2.3
bayesian-optimization==1.0.1
pandas==0.25.1
matplotlib==2.2.2
seaborn==0.9.0
xgboost==0.90
numpy==1.17.0
chinesecalendar==1.2.2
scikit-learn==0.21.3
torch==1.3.1
torchvision==0.4.2
statsmodels==0.10.1
pyecharts==0.5.10
pyecharts-snapshot==0.1.10
python-docx==0.8.10
pycrypto==2.6.1


